requirements.txt
docopt==0.6.1
mock==1.0.1
py==1.4.20
pytest==2.5.2
mplayer.py
cx_Freeze
